Update Tracking.md
I recommend switching the order of installation of YOLO and pytorch.  I believe that pytorch updates the cuda libraries and this appears to have caused some issues during our installations.  I added a few comments to sections "install pytorch" and "download weights" as well as an intermediary test step to make sure darknet has installed correctly.  For us, this was one of the major trials due to to variations in CUDA libraries and making sure darknet found the appropriate folders.